+++ GaluhadPlugins/Pets/Strings/French/Pets.lua
@@ -587,6 +587,10 @@ _PETSTRINGS[497] = {[1]="Hibou grand-duc du soleil fleurissant";[2]="Disponible 
 
 _PETSTRINGS[498] = {[1]="Hamster moucheté";[2]="Disponibilité limitée dans les Coffres à butin de voyageur et d'aventurier";}
 _PETSTRINGS[499] = {[1]="Vautour tacheté";[2]="Offre à durée limitée : suivi de récompense d’objets légendaires de la saison 19."}
+_PETSTRINGS[500] = {[1]="Limace rouille";[2]="";}
+_PETSTRINGS[501] = {[1]="Limace nocturne";[2]="";}
+_PETSTRINGS[502] = {[1]="Calf of High-land Pastures";[2]="Available through barter during the Farmers Faire Festival."}
+_PETSTRINGS[503] = {[1]="Protecteur grodbog luxuriant";[2]=""}
 
 _PETSTRINGS[1000] = {[1]="Chauve-souris rayée";[2]=""}
 _PETSTRINGS[1001] = {[1]="Chauve-souris brunie";[2]=""}
@@ -608,38 +612,35 @@ _PETSTRINGS[1016] = {[1]="Terreur des cavernes";[2]="";}
 _PETSTRINGS[1017] = {[1]="Jeune griffefosse verdoyant";[2]=""}
 _PETSTRINGS[1019] = {[1]="Protecteur grodbog venimeux";[2]=""}
 _PETSTRINGS[1020] = {[1]="Reine des grodbogs agressive";[2]=""}
-_PETSTRINGS[1021] = {[1]="Protecteur grodbog luxuriant";[2]=""}
-_PETSTRINGS[1022] = {[1]="Scorpion sanglant";[2]=""}
-_PETSTRINGS[1023] = {[1]="Scorpion noir comme la nuit";[2]=""}
-_PETSTRINGS[1024] = {[1]="Jeune griffefosse pâle";[2]="";}
-_PETSTRINGS[1025] = {[1]="Reine des grodbogs crépusculaire";[2]="";}
-_PETSTRINGS[1026] = {[1]="Scorpion pâle";[2]="";}
-_PETSTRINGS[1027] = {[1]="Protecteur grodbog lumineux";[2]="";}
-_PETSTRINGS[1028] = {[1]="Protecteur grodbog rayé";[2]=""}
-_PETSTRINGS[1029] = {[1]="Reine des grodbogs sombre";[2]=""}
-_PETSTRINGS[1030] = {[1]="Jeune griffefosse rayé";[2]=""}
-_PETSTRINGS[1031] = {[1]="Chien de berger auburn de la Mûr";[2]=""}
-_PETSTRINGS[1032] = {[1]="Chien de berger caramel de la Mûr";[2]=""}
-_PETSTRINGS[1033] = {[1]="Limace du désert";[2]=""}
-_PETSTRINGS[1034] = {[1]="Guêpe du désert en armure";[2]="";}
-_PETSTRINGS[1035] = {[1]="Ver endurci par les combats";[2]="";}
-_PETSTRINGS[1036] = {[1]="Bébé mûmak brun";[2]="";}
-_PETSTRINGS[1037] = {[1]="Bébé mûmak assidu";[2]=""}
-_PETSTRINGS[1038] = {[1]="Zizanu ombré";[2]=""}
-_PETSTRINGS[1039] = {[1]="Ver rusé";[2]=""}
-_PETSTRINGS[1040] = {[1]="Bébé mûmak travailleur";[2]=""}
-_PETSTRINGS[1041] = {[1]="Bébé mûmak marbré";[2]=""}
-_PETSTRINGS[1042] = {[1]="Bébé mûmak illuminé";[2]=""}
-_PETSTRINGS[1043] = {[1]="Ver rôdeur";[2]=""}
-_PETSTRINGS[1044] = {[1]="Limace maritime";[2]="";}
-_PETSTRINGS[1045] = {[1]="Zizanu cuivré";[2]="";}
-_PETSTRINGS[1046] = {[1]="Bébé mûmak du crépuscule";[2]="";}
-_PETSTRINGS[1047] = {[1]="Limace nocturne";[2]=""}
-_PETSTRINGS[1048] = {[1]="Limace rouille";[2]=""}
-_PETSTRINGS[1049] = {[1]="Bébé mûmak surchargé";[2]=""}
-_PETSTRINGS[1050] = {[1]="Jeune crocodile des marais";[2]=""}
-_PETSTRINGS[1051] = {[1]="Bébé mûmak pierreux";[2]=""}
-_PETSTRINGS[1052] = {[1]="Limace émeraude";[2]=""}
-_PETSTRINGS[1053] = {[1]="Ver rubis";[2]=""}
-_PETSTRINGS[1054] = {[1]="Zizanu pâle";[2]="";}
-_PETSTRINGS[1055] = {[1]="Limace framboise";[2]="";}
+_PETSTRINGS[1021] = {[1]="Scorpion sanglant";[2]=""}
+_PETSTRINGS[1022] = {[1]="Scorpion noir comme la nuit";[2]=""}
+_PETSTRINGS[1023] = {[1]="Jeune griffefosse pâle";[2]=""}
+_PETSTRINGS[1024] = {[1]="Reine des grodbogs crépusculaire";[2]="";}
+_PETSTRINGS[1025] = {[1]="Scorpion pâle";[2]="";}
+_PETSTRINGS[1026] = {[1]="Protecteur grodbog lumineux";[2]="";}
+_PETSTRINGS[1027] = {[1]="Protecteur grodbog rayé";[2]="";}
+_PETSTRINGS[1028] = {[1]="Reine des grodbogs sombre";[2]=""}
+_PETSTRINGS[1029] = {[1]="Jeune griffefosse rayé";[2]=""}
+_PETSTRINGS[1030] = {[1]="Chien de berger auburn de la Mûr";[2]=""}
+_PETSTRINGS[1031] = {[1]="Chien de berger caramel de la Mûr";[2]=""}
+_PETSTRINGS[1032] = {[1]="Limace du désert";[2]=""}
+_PETSTRINGS[1033] = {[1]="Guêpe du désert en armure";[2]=""}
+_PETSTRINGS[1034] = {[1]="Ver endurci par les combats";[2]="";}
+_PETSTRINGS[1035] = {[1]="Bébé mûmak brun";[2]="";}
+_PETSTRINGS[1036] = {[1]="Bébé mûmak assidu";[2]="";}
+_PETSTRINGS[1037] = {[1]="Zizanu ombré";[2]=""}
+_PETSTRINGS[1038] = {[1]="Ver rusé";[2]=""}
+_PETSTRINGS[1039] = {[1]="Bébé mûmak travailleur";[2]=""}
+_PETSTRINGS[1040] = {[1]="Bébé mûmak marbré";[2]=""}
+_PETSTRINGS[1041] = {[1]="Bébé mûmak illuminé";[2]=""}
+_PETSTRINGS[1042] = {[1]="Ver rôdeur";[2]=""}
+_PETSTRINGS[1043] = {[1]="Limace maritime";[2]=""}
+_PETSTRINGS[1044] = {[1]="Zizanu cuivré";[2]="";}
+_PETSTRINGS[1045] = {[1]="Bébé mûmak du crépuscule";[2]="";}
+_PETSTRINGS[1046] = {[1]="Bébé mûmak surchargé";[2]="";}
+_PETSTRINGS[1047] = {[1]="Jeune crocodile des marais";[2]=""}
+_PETSTRINGS[1048] = {[1]="Bébé mûmak pierreux";[2]=""}
+_PETSTRINGS[1049] = {[1]="Limace émeraude";[2]=""}
+_PETSTRINGS[1050] = {[1]="Ver rubis";[2]=""}
+_PETSTRINGS[1051] = {[1]="Zizanu pâle";[2]=""}
+_PETSTRINGS[1052] = {[1]="Limace framboise";[2]=""}